requirements.txt
numpy~=1.21.2
fastapi~=0.70.0
uvicorn~=0.15.0
httpx~=0.20.0
websockets~=10.0
pydantic~=1.8.2
python-multipart
aiohttp~=3.7.4.post0
psutil~=5.8.0
starlette~=0.16.0
opencv-python~=4.5.3.56
colorama~=0.4.4
websocket-client~=1.2.1
requests~=2.26.0
pip~=21.3